+++ Windows/Deskestro.Windows.TestApp/GesturePanel.cs
@@ -1,0 +1,9 @@
+namespace Deskestro.Windows.TestApp;
+
+internal sealed class GesturePanel : Panel
+{
+    internal GesturePanel()
+    {
+        SetStyle(ControlStyles.StandardClick | ControlStyles.StandardDoubleClick, true);
+    }
+}