+++ Windows/Deskestro.Windows/AutomationGlobalUsing.cs
@@ -1,0 +1,1 @@
+global using Deskestro.Windows.Automation;